Move lnurl-pay tests from sdk-common to SDK
